fixes to binning in GUI
Former-commit-id: e57e4ac8fb8e03aa44745c063ae42e60605980a8

+++ py4DSTEM/process/preprocess/preprocess.py
@@ -110,7 +110,7 @@ def bin_data_mmap(datacube, bin_factor):
     assert type(bin_factor) is int, "Error: binning factor {} is not an int.".format(bin_factor)
     R_Nx,R_Ny,Q_Nx,Q_Ny = datacube.R_Nx,datacube.R_Ny,datacube.Q_Nx,datacube.Q_Ny
 
-    data = np.zeros((datacube.R_Nx,datacube.R_Ny,datacube.Q_Nx//bin_factor,datacube.Q_Ny//bin_factor),dtype='uint32')
+    data = np.zeros((datacube.R_Nx,datacube.R_Ny,datacube.Q_Nx//bin_factor,datacube.Q_Ny//bin_factor),dtype=np.float64)
     for Rx in range(datacube.R_Nx):
         for Ry in range(datacube.R_Ny):
             data[Rx,Ry,:,:] = bin2D(datacube.data[Rx,Ry,:,:],bin_factor,dtype=np.float64)
